Include PhaseSpaceActor in digitizers.py
